+++ R/adjust_em.R
@@ -78,11 +78,15 @@ adjust_em_val <- function(
   )
 
   x_mod_coefs <- coef(x_mod)
-  x_pred <- x_mod_coefs[1]
+  x_mod_se <- sqrt(diag(vcov(x_mod)))
 
-  for (i in 2:length(x_mod_coefs)) {
+  # Sample coefficients independently using their standard errors
+  x_mod_coefs_sampled <- rnorm(length(x_mod_coefs), x_mod_coefs, x_mod_se)
+  x_pred <- x_mod_coefs_sampled[1]
+
+  for (i in 2:length(x_mod_coefs_sampled)) {
     var_name <- names(x_mod_coefs)[i]
-    x_pred <- x_pred + df[[var_name]] * x_mod_coefs[i]
+    x_pred <- x_pred + df[[var_name]] * x_mod_coefs_sampled[i]
   }
 
   df$Xpred <- rbinom(n, 1, plogis(x_pred))