+++ R/multibias_adjust.R
@@ -31,9 +31,9 @@
 #' @param level Value from 0-1 representing the full range of the confidence
 #' interval. Default is 0.95.
 #'
-#' @return A list where the first item is the odds ratio estimate of the
-#' effect of the exposure on the outcome and the second item is the
-#' confidence interval as the vector: (lower bound, upper bound).
+#' @return A list including: the bias-adjusted effect estimate of the exposure
+#' on the outcome, the standard error, and the confidence interval as the
+#' vector: (lower bound, upper bound).
 #'
 #' @examples
 #' # Adjust for exposure misclassification -------------------------------------
@@ -127,7 +127,7 @@
 #' @import dplyr
 #' @importFrom magrittr %>%
 #' @importFrom stats as.formula binomial coef lm median glm plogis rbinom
-#' @importFrom stats qnorm quantile
+#' @importFrom stats rnorm sd qnorm quantile vcov
 #' @importFrom rlang .data
 #'
 #' @export
@@ -185,6 +185,7 @@ multibias_adjust <- function(
     alpha <- 1 - level
     output <- list(
       estimate = median(est),
+      std.error = sd(est),
       ci = as.vector(quantile(est, c(alpha / 2, 1 - alpha / 2)))
     )
   }