+++ tests/testthat/test-adjust_om_sel.R
@@ -35,7 +35,7 @@ list_for_om_sel <- list(
 )
 bp_om_sel <- bias_params(coef_list = list_for_om_sel)
 
-single_run <- adjust_om_sel(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_om_sel
 )
@@ -91,7 +91,7 @@ list_for_om_sel <- list(
 )
 bp_om_sel <- bias_params(coef_list = list_for_om_sel)
 
-single_run <- adjust_om_sel(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_om_sel
 )
@@ -115,7 +115,7 @@ test_that("odds ratio and confidence interval output", {
 
 # adjust with validation data
 
-or_val <- adjust_om_sel(
+val_run <- multibias_adjust(
   data_observed = data_observed(
     df_om_sel,
     bias = c("om", "sel"),
@@ -130,10 +130,12 @@ or_val <- adjust_om_sel(
     confounders = c("C1", "C2", "C3"),
     misclassified_outcome = "Ystar",
     selection = "S"
-  )
+  ),
+  bootstrap = TRUE,
+  bootstrap_reps = nreps
 )
 
 test_that("adjust_om_sel, validation data", {
-  expect_gt(or_val$estimate, or_true - 0.1)
-  expect_lt(or_val$estimate, or_true + 0.1)
+  expect_gt(val_run$estimate, or_true - 0.1)
+  expect_lt(val_run$estimate, or_true + 0.1)
 })