+++ tests/testthat/test-adjust_uc_em.R
@@ -35,7 +35,7 @@ list_for_uc_em <- list(
 )
 bp_uc_em <- bias_params(coef_list = list_for_uc_em)
 
-single_run <- adjust_uc_em(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_em
 )
@@ -90,7 +90,7 @@ list_for_uc_em <- list(
 )
 bp_uc_em <- bias_params(coef_list = list_for_uc_em)
 
-single_run <- adjust_uc_em(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_em
 )
@@ -143,7 +143,7 @@ list_for_uc_em <- list(
 )
 bp_uc_em <- bias_params(coef_list = list_for_uc_em)
 
-single_run <- adjust_uc_em(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_em
 )
@@ -195,7 +195,7 @@ list_for_uc_em <- list(
 )
 bp_uc_em <- bias_params(coef_list = list_for_uc_em)
 
-single_run <- adjust_uc_em(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_em
 )
@@ -219,7 +219,7 @@ test_that("odds ratio and confidence interval output", {
 
 # adjust with validation data
 
-or_val <- adjust_uc_em(
+val_run <- multibias_adjust(
   data_observed = data_observed(
     df_uc_em,
     bias = c("uc", "em"),
@@ -233,10 +233,12 @@ or_val <- adjust_uc_em(
     true_outcome = "Y",
     confounders = c("C1", "C2", "C3", "U"),
     misclassified_exposure = "Xstar"
-  )
+  ),
+  bootstrap = TRUE,
+  bootstrap_reps = nreps
 )
 
 test_that("adjust_uc_em, validation data", {
-  expect_gt(or_val$estimate, or_true - 0.1)
-  expect_lt(or_val$estimate, or_true + 0.1)
+  expect_gt(val_run$estimate, or_true - 0.1)
+  expect_lt(val_run$estimate, or_true + 0.1)
 })