+++ tests/testthat/test-adjust_uc_om.R
@@ -31,7 +31,7 @@ list_for_uc_om <- list(
 )
 bp_uc_om <- bias_params(coef_list = list_for_uc_om)
 
-single_run <- adjust_uc_om(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_om
 )
@@ -82,7 +82,7 @@ list_for_uc_om <- list(
 )
 bp_uc_om <- bias_params(coef_list = list_for_uc_om)
 
-single_run <- adjust_uc_om(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_om
 )
@@ -136,7 +136,7 @@ list_for_uc_om <- list(
 )
 bp_uc_om <- bias_params(coef_list = list_for_uc_om)
 
-single_run <- adjust_uc_om(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_om
 )
@@ -188,7 +188,7 @@ list_for_uc_om <- list(
 )
 bp_uc_om <- bias_params(coef_list = list_for_uc_om)
 
-single_run <- adjust_uc_om(
+single_run <- multibias_adjust(
   df_observed,
   bias_params = bp_uc_om
 )
@@ -212,7 +212,7 @@ test_that("odds ratio and confidence interval output", {
 
 # adjust with validation data
 
-or_val <- adjust_uc_om(
+val_run <- multibias_adjust(
   data_observed = data_observed(
     df_uc_om,
     bias = c("uc", "om"),
@@ -226,10 +226,12 @@ or_val <- adjust_uc_om(
     true_outcome = "Y",
     confounders = c("C1", "C2", "C3", "U"),
     misclassified_outcome = "Ystar"
-  )
+  ),
+  bootstrap = TRUE,
+  bootstrap_reps = nreps
 )
 
 test_that("adjust_uc_om, validation data", {
-  expect_gt(or_val$estimate, or_true - 0.1)
-  expect_lt(or_val$estimate, or_true + 0.1)
+  expect_gt(val_run$estimate, or_true - 0.1)
+  expect_lt(val_run$estimate, or_true + 0.1)
 })